Refactor stories and tests.
